+++ cnd/src/db.rs
@@ -25,7 +25,8 @@ pub use self::{
 
 use crate::{
     db::wrapper_types::custom_sql_types::Text,
-    swap_protocols::{rfc003::SwapId, LocalSwapId, Role},
+    swap_protocols::{rfc003::SwapId, Role},
+    LocalSwapId,
 };
 use async_trait::async_trait;
 use comit::Protocol;
