+++ cnd/src/db/tables.rs
@@ -9,7 +9,8 @@ use crate::{
         Sqlite,
     },
     identity, lightning,
-    swap_protocols::{halight, hbit, herc20, rfc003, LocalSwapId, Role, Side},
+    swap_protocols::{halight, hbit, herc20, rfc003, Role, Side},
+    LocalSwapId,
 };
 use anyhow::Context;
 use diesel::{prelude::*, RunQueryDsl};
